# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Glotelho API Documentation",
        "description": "API de raccourcissement d'URL avec suivi des statistiques",
        "contact": {
            "email": "support@glotelho.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/shorten": {
            "post": {
                "tags": [
                    "URLs"
                ],
                "summary": "Raccourcir une URL",
                "description": "Raccourcir une URL",
                "operationId": "392a6c23593d7916a29398bfb5df61f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://exemple.com/very/long/url"
                                    },
                                    "custom_code": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "minLength": 3,
                                        "example": "mon-site"
                                    },
                                    "expires_in": {
                                        "description": "Nombre de jours avant expiration",
                                        "type": "integer",
                                        "maximum": 365,
                                        "minimum": 1,
                                        "example": 30
                                    },
                                    "expires_at": {
                                        "description": "Date d'expiration spécifique",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-31T23:59:59Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "URL raccourcie avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "URL raccourcie avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ShortUrl"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/check/{code}": {
            "get": {
                "tags": [
                    "URLs"
                ],
                "summary": "Vérifier la disponibilité d'un code personnalisé",
                "description": "Vérifier la disponibilité d'un code personnalisé",
                "operationId": "52f2f3b4f72d42e2dbd89d8fa766cf1c",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Le code personnalisé à vérifier",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "mon-site"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Code disponible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "available": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Code déjà utilisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "available": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ce code est déjà utilisé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ShortUrl": {
                "properties": {
                    "short_url": {
                        "type": "string",
                        "example": "http://localhost/abc123"
                    },
                    "original_url": {
                        "type": "string",
                        "example": "https://exemple.com/very/long/url"
                    },
                    "short_code": {
                        "type": "string",
                        "example": "abc123"
                    },
                    "is_custom": {
                        "type": "boolean",
                        "example": false
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-31T23:59:59Z",
                        "nullable": true
                    },
                    "click_count": {
                        "type": "integer",
                        "example": 42
                    }
                },
                "type": "object"
            },
            "Error": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Message d'erreur détaillé"
                    }
                },
                "type": "object"
            }
        }
    },
...
}